refactor code and fix app description
