+++ app/ai/openai-client.js
