+++ config/.env.default
@@ -1,3 +1,3 @@
-# OpenRouter API key - required for AI summaries
+# OpenRouter API key - required only when using AI model from OpenRouter
 # Get your API key from: https://openrouter.ai/keys
 OPENROUTER_API_KEY=
